fix: Fix const constructor and improve type conversions
Fixed remaining compilation errors:

1. ai_insight.dart:51 - Removed 'const' from constructor
   since DateTime.now() is not a const expression

2-3. analysis_screen.dart:205-206 - Changed type conversion approach
   from: 90 + (position.dx * 80).toDouble()
   to: 90.0 + position.dx * 80
   Using double literal ensures proper type inference

4-5. scoring_screen.dart:238,275 - Use arrows.length directly
   instead of arrowCount.toInt() to avoid num type issues

6-7. scoring_screen.dart:292 - Changed type conversion approach
   from: 160 + (position.dy * 140).toDouble()
   to: 160.0 + position.dy * 140
   Using double literal ensures proper type inference

All changes ensure proper type compatibility for iOS build.

+++ lib/screens/analysis_screen.dart
@@ -202,8 +202,8 @@ class AnalysisScreen extends ConsumerWidget {
               if (hasData)
                 ...stats.heatmapData.take(30).map((position) {
                   return Positioned(
-                    left: 90 + (position.dx * 80).toDouble(),
-                    top: 90 + (position.dy * 80).toDouble(),
+                    left: 90.0 + position.dx * 80,
+                    top: 90.0 + position.dy * 80,
                     child: Container(
                       width: 8,
                       height: 8,
